DeepVision/requirements.txt
alembic==1.4.1
blis==0.7.5
Boruta==0.3
brotlipy==0.7.0
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
colorlover==0.3.0
cryptography==36.0.0
cufflinks==0.17.3
cymem==2.0.6
Cython==0.29.14
databricks-cli==0.16.2
distlib==0.3.4
docker==5.0.3
filelock==3.4.2
Flask==2.0.3
Flask-Cors==3.0.10
funcy==1.17
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.25
greenlet==1.1.2
identify==2.4.2
idna==3.3
image-slicer==2.1.1
ImageHash==4.2.1
imbalanced-learn==0.7.0
ipywidgets==7.6.5
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.0.1
jupyterlab-widgets==1.0.2
kmodes==0.11.1
ldap3==2.9.1
lightgbm==3.3.2
Mako==1.1.6
missingno==0.5.0
mlflow==1.22.0
mlxtend==0.19.0
multimethod==1.6
murmurhash==1.0.6
networkx==2.6.3
nltk==3.6.7
nodeenv==1.6.0
numba==0.53.0
numexpr==2.8.1
pandas==1.3.5
pandas-profiling==3.1.0
patsy==0.5.2
phik==0.12.0
pip==21.2.2
plac==1.1.3
platformdirs==2.4.1
plotly==5.5.0
pre-commit==2.16.0
preshed==3.0.6
prometheus-flask-exporter==0.18.7
pyasn1==0.4.8
pycaret==2.3.5
pycparser==2.21
pydantic==1.9.0
pydirectory==0.1.15
pyLDAvis==3.2.2
pynndescent==0.5.5
pyod==0.9.7
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
querystring-parser==1.2.4
regex==2021.11.10
requests==2.27.1
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
setuptools==58.0.4
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==2.3.7
SQLAlchemy==1.4.29
sqlparse==0.4.2
srsly==1.0.5
statsmodels==0.12.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
textblob==0.17.1
thinc==7.4.5
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.8
virtualenv==20.13.0
visions==0.7.4
waitress==2.0.0
wasabi==0.9.0
websocket-client==1.2.3
Werkzeug==2.0.3
wheel==0.37.0
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.1
yellowbrick==1.3.post1
